+++ extension/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to the "doclin" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.1.4]
+
+### Changed
+- Updated thread addition behaviour under "File Threads" to attach to code file without snippet.
+- Redesigned UI.
+
+### Added
+- Added title to threads.
+
 ## [0.1.3]
 
 ### Changed